==> scripts/08_FARMS.R <==
#' -------------------------------------
#' Look at FARM counts
#' across berms versus controls and
#' -------------------------------------

source("scripts/helpers.R")


#load libraries
library(tidyverse)
library(broom)
library(lubridate)
library(glmmTMB)
library(performance)
library(car)
library(broom.mixed)

# load data
farm_dat <- read_csv("data/farms.csv")

# Work it up
farm_dat <- farm_dat |> 
  mutate(date_deployed = make_date(year_deployed, month_deployed, day_deployed)) |>
  mutate(date_retrieved = make_date(year_retrieved, month_retrieved, day_retrieved)) |>
  mutate(soak_duration = date_retrieved - date_deployed) |> 
  select(-c(4:14,16:20)) |> 
  select(1,2,3,5,6,7,4) |> 
  group_by(site, treatment, farms_number, date_deployed, date_retrieved, soak_duration, species_code) |>
  reframe(abundance = n()) |> 
  pivot_wider(names_from = species_code, values_from = abundance, values_fill = 0) |> 
  filter(!is.na(date_retrieved)) |> 
  select(-c(NOSP, LESP, CRFO, ULSL, LILI)) |> # get rid of nothing placeholder, ULSL is supposed to be ULSP which is an algae, not sure why it was recorded, CRFO only recoreded on one (i.e. not consistently recorded), ditto for lesp. Also droppuing LILI, not assessed in winter I dont think
  mutate(soak_class = if_else(condition = soak_duration > 100, "long", "short")) |> 
  rename(FUNS = FUSP) # correcting this code. FUSP is an algae, FUNS is a fish


# Calc abundance found per farm by retrieval season/soak duration
farm_soak_dat <- farm_dat |>
  select(1,2,19, 7:18) 

farm_soak_dat <- farm_soak_dat |> 
  group_by(site, treatment, soak_class) |> 
  summarise(
    n_farms = n(),
    across(
      where(is.numeric),
      \(x) sum(x, na.rm = TRUE)), # sum up across farms
    .groups = "drop"
  ) |>
  mutate(
    across(
      -c(1:4),
      ~ round(.x / n_farms,2))) # standardize by count per farm

  
  
# FPEN, CAMA, LILI, PHGU, PAHE are the codes that occur in abundance
# LILI exclusively in the short/summer soak
# FPEN most abundant in the long soak, but still present in the short soak
# PAHE only in winter/long soak. Seasonality or duration thing?
# PHGU mostly in winter/long soak. Seasonality or duration thing?

# For this analysis, I will focus only on those spp which occur at some abundance (FPEN, CAMA, LILI, PHGU, PAHE)
# Also excluding coughlin due to not enough farms recovered

# CAMA
CAMA_mod <- glmmTMB(CAMA ~ treatment + soak_class, 
                     data = farm_dat |> filter(site == "Bayside"), 
                     family = poisson(link = "log"))

# check_model(CAMA_mod)
# 
# summarizing tests
Anova(CAMA_mod) |> tidy() # Soak class seems to explain more than the treatment
tidy(CAMA_mod)

CAMA_plot <- ggplot(farm_dat |> filter(site == "Bayside"), 
                    mapping = aes(x = treatment, 
                                  y = CAMA,
                                  fill = treatment)) +
  geom_boxplot() +
  geom_jitter(aes(shape = soak_class),  
              width = 0.2,
              height = 0, 
              alpha = 0.6, 
              size = 2) +
  labs(title = "CAMA abundance in farms",
       x = "Treatment",
       y = "CAMA Abundance") +
  facet_wrap(vars(soak_class))



# FPEN
FPEN_mod <- glmmTMB(FPEN ~ treatment + soak_class, 
                    data = farm_dat |> filter(site == "Bayside"), 
                    family = poisson(link = "log"))

# check_model(FPEN_mod)
# 
# # summarizing tests
# Anova(FPEN_mod) |> tidy() 
# tidy(FPEN_mod)
# Fewer shrimps in the control and in the short soak time. Excluded coughlin due to low soak time

ggplot(farm_dat |> filter(site == "Bayside"), 
       mapping = aes(x = treatment, 
                     y = FPEN,
                     fill = treatment)) +
  geom_boxplot() +
  geom_jitter(aes(shape = soak_class),  
              width = 0.2,
              height = 0, 
              alpha = 0.6, 
              size = 2) +
  labs(title = "FPEN abundance in farms",
       x = "Treatment",
       y = "FPEN Abundance") +
  facet_wrap(vars(soak_class))


# LILI - Ignoring LILI here. they were measured in other ways, and I'm not sure that they were reliably counted in the winter time point
# LILI_mod <- glmmTMB(LILI ~ treatment + site, 
#                     data = farm_dat |> filter(soak_class == "short"), 
#                     family = poisson(link = "log"))
# 
# # check_model(LILI_mod)
# # 
# # # summarizing tests
# Anova(LILI_mod) |> tidy()
# tidy(LILI_mod)
# 
# # no LILI in the long/winter soak class. Coughlin seems to support more than Bayside, no effect of treatment
# # I'm concerned we may not have counted LILI in the winter retrieval but i'm not sure
# ggplot(farm_dat |> filter(soak_class == "short"), 
#        mapping = aes(x = treatment, 
#                      y = LILI,
#                      fill = treatment)) +
#   geom_boxplot() +
#   geom_jitter(aes(shape = site),  
#               width = 0.2,
#               height = 0, 
#               alpha = 0.6, 
#               size = 2) +
#   labs(title = "LILI abundance in farms after a short soak",
#        x = "Treatment",
#        y = "LILI Abundance") +
#   facet_wrap(vars(site))

# PHGU - Bayside only
PHGU_mod <- glmmTMB(PHGU ~ treatment + soak_class, 
                    data = farm_dat |> filter(site == "Bayside"), 
                    family = poisson(link = "log"))



# check_model(PHGU_mod)
# 
# # summarizing tests
# Anova(PHGU_mod) |> tidy() 
# tidy(PHGU_mod)
# Lose PHGU in the control and the short


ggplot(farm_dat |> filter(site == "Bayside"), 
       mapping = aes(x = treatment, 
                     y = PHGU,
                     fill = treatment)) +
  geom_boxplot() +
  geom_jitter(aes(shape = soak_class),  
              width = 0.2,
              height = 0, 
              alpha = 0.6, 
              size = 2) +
  labs(title = "PHGU abundance in farms",
       x = "Treatment",
       y = "PHGU Abundance") +
  facet_wrap(vars(soak_class))
